+++ pepdbagent/modules/project.py
@@ -8,12 +8,14 @@ from sqlalchemy import and_, delete, select
 from sqlalchemy.exc import IntegrityError, NoResultFound
 from sqlalchemy.orm import Session
 from sqlalchemy import Select
+import numpy as np
 
 from peppy.const import (
     SAMPLE_RAW_DICT_KEY,
     SUBSAMPLE_RAW_LIST_KEY,
     CONFIG_KEY,
     SAMPLE_TABLE_INDEX_KEY,
+    SAMPLE_NAME_ATTR,
 )
 
 from pepdbagent.const import (
@@ -24,8 +26,12 @@ from pepdbagent.const import (
 )
 
 from pepdbagent.db_utils import Projects, Samples, Subsamples, BaseEngine
-from pepdbagent.exceptions import ProjectNotFoundError, ProjectUniqueNameError
-from pepdbagent.models import UpdateItems, UpdateModel
+from pepdbagent.exceptions import (
+    ProjectNotFoundError,
+    ProjectUniqueNameError,
+    PEPDatabaseAgentError,
+)
+from pepdbagent.models import UpdateItems, UpdateModel, ProjectDict
 from pepdbagent.utils import create_digest, registry_path_converter
 
 
@@ -210,7 +216,7 @@ class PEPDatabaseProject:
 
     def create(
         self,
-        project: peppy.Project,
+        project: Union[peppy.Project, dict],
         namespace: str,
         name: str = None,
         tag: str = DEFAULT_TAG,
@@ -227,6 +233,13 @@ class PEPDatabaseProject:
         update is set True)
 
         :param peppy.Project project: Project object that has to be uploaded to the DB
+            danger zone:
+                optionally, project can be a dictionary with PEP elements
+                ({
+                    _config: dict,
+                    _sample_dict: Union[list, dict],
+                    _subsample_list: list
+                })
         :param namespace: namespace of the project (Default: 'other')
         :param name: name of the project (Default: name is taken from the project object)
         :param tag: tag (or version) of the project.
@@ -239,9 +252,22 @@ class PEPDatabaseProject:
         :param description: description of the project
         :return: None
         """
-        proj_dict = project.to_dict(extended=True, orient="records")
+        if isinstance(project, peppy.Project):
+            proj_dict = project.to_dict(extended=True, orient="records")
+        elif isinstance(project, dict):
+            # verify if the dictionary has all necessary elements.
+            # samples should be always presented as list of dicts (orient="records"))
+            _LOGGER.warning(
+                f"Project f{namespace}/{name}:{tag} is provided as dictionary. Project won't be validated."
+            )
+            proj_dict = ProjectDict(**project).model_dump(by_alias=True)
+        else:
+            raise PEPDatabaseAgentError(
+                "Project has to be peppy.Project object or dictionary with PEP elements"
+            )
+
         if not description:
-            description = project.description
+            description = project.get(description, "")
         proj_dict[CONFIG_KEY][DESCRIPTION_KEY] = description
 
         namespace = namespace.lower()
@@ -255,7 +281,10 @@ class PEPDatabaseProject:
         proj_dict[CONFIG_KEY][NAME_KEY] = proj_name
 
         proj_digest = create_digest(proj_dict)
-        number_of_samples = len(project.samples)
+        try:
+            number_of_samples = len(project.samples)
+        except AttributeError:
+            number_of_samples = len(proj_dict[SAMPLE_RAW_DICT_KEY])
 
         if update_only:
             _LOGGER.info(f"Update_only argument is set True. Updating project {proj_name} ...")
@@ -293,7 +322,9 @@ class PEPDatabaseProject:
                 self._add_samples_to_project(
                     new_prj,
                     proj_dict[SAMPLE_RAW_DICT_KEY],
-                    sample_table_index=project.sample_table_index,
+                    sample_table_index=proj_dict[CONFIG_KEY].get(
+                        SAMPLE_TABLE_INDEX_KEY, SAMPLE_NAME_ATTR
+                    ),
                 )
 
                 if proj_dict[SUBSAMPLE_RAW_LIST_KEY]:
@@ -833,6 +864,7 @@ class PEPDatabaseProject:
                 namespace=original_namespace,
                 name=original_name,
                 tag=original_tag,
+                raw=True,
             ),
             namespace=fork_namespace,
             name=fork_name,
@@ -857,6 +889,80 @@ class PEPDatabaseProject:
             fork_prj.forked_from_id = original_prj.id
             fork_prj.pop = original_prj.pop
             fork_prj.submission_date = original_prj.submission_date
+            fork_prj.pep_schema = original_prj.pep_schema
+            fork_prj.description = description or original_prj.description
 
             session.commit()
         return None
+
+    def get_config(self, namespace: str, name: str, tag: str) -> Union[dict, None]:
+        """
+        Get project configuration by providing namespace, name, and tag
+
+        :param namespace: project namespace
+        :param name: project name
+        :param tag: project tag
+        :return: project configuration
+        """
+        statement = select(Projects.config).where(
+            and_(Projects.namespace == namespace, Projects.name == name, Projects.tag == tag)
+        )
+        with Session(self._sa_engine) as session:
+            result = session.execute(statement).one_or_none()
+
+        if result:
+            return result[0]
+        return None
+
+    def get_subsamples(self, namespace: str, name: str, tag: str) -> Union[list, None]:
+        """
+        Get project subsamples by providing namespace, name, and tag
+
+        :param namespace: project namespace
+        :param name: project name
+        :param tag: project tag
+        :return: list with project subsamples
+        """
+        statement = self._create_select_statement(name, namespace, tag)
+
+        with Session(self._sa_engine) as session:
+
+            found_prj = session.scalar(statement)
+
+            if found_prj:
+                _LOGGER.info(f"Project has been found: {found_prj.namespace}, {found_prj.name}")
+                subsample_dict = {}
+                if found_prj.subsamples_mapping:
+                    for subsample in found_prj.subsamples_mapping:
+                        if subsample.subsample_number not in subsample_dict.keys():
+                            subsample_dict[subsample.subsample_number] = []
+                        subsample_dict[subsample.subsample_number].append(subsample.subsample)
+                    return list(subsample_dict.values())
+                else:
+                    return []
+            else:
+                raise ProjectNotFoundError(
+                    f"No project found for supplied input: '{namespace}/{name}:{tag}'. "
+                    f"Did you supply a valid namespace and project?"
+                )
+
+    def get_samples(self, namespace: str, name: str, tag: str, raw: bool = True) -> list:
+        """
+        Get project samples by providing namespace, name, and tag
+
+        :param namespace: project namespace
+        :param name: project name
+        :param tag: project tag
+        :param raw: if True, retrieve unprocessed (raw) PEP dict. [Default: True]
+
+        :return: list with project samples
+        """
+        if raw:
+            return self.get(namespace=namespace, name=name, tag=tag, raw=True).get(
+                SAMPLE_RAW_DICT_KEY
+            )
+        return (
+            self.get(namespace=namespace, name=name, tag=tag, raw=False)
+            .sample_table.replace({np.nan: None})
+            .to_dict(orient="records")
+        )